fixup! feat(ram): Add flat Elf loading capabilities
