+++ src/test/csrc/common/elfloader.cpp
@@ -15,6 +15,17 @@
 
 #include "elfloader.h"
 
+void ElfBinary::load() {
+  assert(size >= sizeof(Elf64_Ehdr));
+  eh64 = (const Elf64_Ehdr *)raw;
+  assert(IS_ELF32(*eh64) || IS_ELF64(*eh64));
+
+  if (IS_ELF32(*eh64))
+    parse(data32);
+  else
+    parse(data64);
+}
+
 template <typename ehdr_t, typename phdr_t, typename shdr_t, typename sym_t>
 void ElfBinary::parse(ElfBinaryData<ehdr_t, phdr_t, shdr_t, sym_t> &data) {
   data.eh = (const ehdr_t *)raw;
@@ -35,6 +46,25 @@ void ElfBinary::parse(ElfBinaryData<ehdr_t, phdr_t, shdr_t, sym_t> &data) {
       }
     }
   }
+}
+
+ElfBinaryFile::ElfBinaryFile(const char *filename) : filename(filename) {
+  int fd = open(filename, O_RDONLY);
+  struct stat s;
+  assert(fd != -1);
+  assert(fstat(fd, &s) >= 0);
+  size = s.st_size;
+
+  raw = (uint8_t *)mmap(NULL, size, PROT_READ, MAP_PRIVATE, fd, 0);
+  assert(raw != MAP_FAILED);
+  close(fd);
+
+  load();
+}
+
+ElfBinaryFile::~ElfBinaryFile() {
+  if (raw)
+    munmap((void *)raw, size);
 }
 
 bool isElfFile(const char *filename) {
@@ -69,6 +99,12 @@ long readFromElf(void *ptr, const char *file_name, long buf_size) {
 
   uint64_t len_written = 0;
   auto base_addr = elf_file.sections[0].data_dst;
+
+  if (base_addr != PMEM_BASE) {
+    printf(
+        "The first address in the elf does not match the base of the physical memory. It is likely that execution "
+        "leads to unexpected behaviour.");
+  }
 
   for (auto section: elf_file.sections) {
     auto len = section.data_len + section.zero_len;
